Restructure endpoints
This changes almost all the endpoints after the following considerations:

`page-id` as part of the URL is better than as GET param

Always having page-id as part of the URL is better than having it only in
some cases.
We propose using a special character '-' instead of page-id when the id is
not needed and was inspired by the Google API Design Patters
https://cloud.google.com/apis/design/design_patterns#list_sub-collections .
This makes the URL format clearer, simpler and it also makes
it easier to write DRF routers.

The moderation/publishing endpoints are now disconnected from specific revisions
and they create new revisions.
This is also how Wagtail works internally so it will be easier to implement.